Move libyasm-stdint.h include out of util.h, as util.h is included from build platform files.
While here, fix a few warnings by pushing uintptr_t to a few more register
usages.

Noticed by: rugxulo@gmail.com

svn path=/trunk/yasm/; revision=1786

+++ libyasm/arch.c
@@ -29,6 +29,7 @@
 #include "util.h"
 /*@unused@*/ RCSID("$Id$");
 
+#include "libyasm-stdint.h"
 #include "coretype.h"
 
 #include "expr.h"
@@ -131,7 +132,8 @@ yasm_operand_print(const yasm_insn_operand *op, FILE *f, int indent_level,
 	    fprintf(f, "\n");
 	    break;
     }
-    fprintf(f, "%*sTargetMod=%lx\n", indent_level+1, "", op->targetmod);
+    fprintf(f, "%*sTargetMod=%lx\n", indent_level+1, "",
+	    (unsigned long)op->targetmod);
     fprintf(f, "%*sSize=%u\n", indent_level+1, "", op->size);
     fprintf(f, "%*sDeref=%d, Strict=%d\n", indent_level+1, "", (int)op->deref,
 	    (int)op->strict);
